% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustRcheck.R
\name{heatmap_visual}
\alias{heatmap_visual}
\title{Visual Check of Clusters for Unimodality}
\usage{
heatmap_visual(data, cluster = "cluster", ...)
}
\arguments{
\item{data}{A `data.frame` object}

\item{cluster}{Specify the cluster column's name}

\item{...}{Extra arguments, not used}
}
\description{
\code{heatmap_visual} builds a heatmap of all clusters where
the results of the test are represented as colors.
}
